photos_app_backend/requirements.txt
django>=3.2,<4.0
djangorestframework
django-cors-headers
pillow
gunicorn